requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Deep learning and transformers
transformers>=4.35.0
sentence-transformers>=2.2.0
accelerate>=0.24.0
safetensors>=0.4.0

# Multimodal models and embeddings
timm>=0.9.0
open_clip_torch>=2.23.0
pillow>=10.0.0

# Data processing
datasets>=2.14.0
h5py>=3.9.0
tqdm>=4.66.0

# Evaluation and metrics
scikit-metrics>=0.1.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration and logging
pyyaml>=6.0
hydra-core>=1.3.0
tensorboard>=2.14.0
wandb>=0.15.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0

# Optional: For LVLM integration
# llava>=1.6.0  # Install separately if needed
# openai>=1.0.0  # For API-based models
